api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.5.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
